requirements.txt
fastapi[standard]
motor
bcrypt
PyJWT